# swagger/restapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "ScyllaDB Management API",
    "version": "0.1"
  },
  "basePath": "/api/v1",
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "repairConfig": {
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean"
        },
        "segment_size_limit": {
          "type": "integer",
          "format": "int64"
        },
        "retry_limit": {
          "type": "integer"
        },
        "retry_backoff_seconds": {
          "type": "integer"
        },
        "parallel_shard_percent": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "repairUnit": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "cluster_id": {
          "type": "string"
        },
        "keyspace": {
          "type": "string"
        },
        "tables": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "repairUnitUpdate": {
      "type": "object",
      "properties": {
        "keyspace": {
          "type": "string"
        },
        "tables": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "repairProgress": {
      "type": "object",
      "properties": {
        "keyspace": {
          "type": "string"
        },
        "tables": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "type": "string"
        },
        "percent_complete": {
          "type": "number",
          "format": "int32"
        },
        "total": {
          "type": "number",
          "format": "int32"
        },
        "success": {
          "type": "number",
          "format": "int32"
        },
        "error": {
          "type": "number",
          "format": "int32"
        },
        "hosts": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "properties": {
              "percent_complete": {
                "type": "number",
                "format": "int32"
              },
              "total": {
                "type": "number",
                "format": "int32"
              },
              "success": {
                "type": "number",
                "format": "int32"
              },
              "error": {
                "type": "number",
                "format": "int32"
              },
              "shards": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "percent_complete": {
                      "type": "number",
                      "format": "int32"
                    },
                    "total": {
                      "type": "number",
                      "format": "int32"
                    },
                    "success": {
                      "type": "number",
                      "format": "int32"
                    },
                    "error": {
                      "type": "number",
                      "format": "int32"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/cluster/{cluster_id}/repair/config": {
      "parameters": [
        {
          "type": "string",
          "description": "cluster ID this API is performing on",
          "name": "cluster_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "description": "getConfig cluster description",
        "responses": {
          "200": {
            "description": "cluster config fields",
            "schema": {
              "$ref": "#/definitions/repairConfig"
            }
          }
        }
      },
      "put": {
        "description": "updateConfig cluster description",
        "parameters": [
          {
            "name": "conf",
            "in": "body",
            "description": "config files",
            "schema": {
              "$ref": "#/definitions/repairConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "deleteConfig cluster description",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/config/{config_type}/{external_id}": {
      "parameters": [
        {
          "type": "string",
          "description": "cluster ID this API is performing on",
          "name": "cluster_id",
          "in": "path",
          "required": true
        },
        {
          "type": "string",
          "description": "unit/keyspace",
          "name": "config_type",
          "in": "path",
          "required": true
        },
        {
          "type": "string",
          "description": "id of the entity being configured",
          "name": "external_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "description": "getConfig cluster description",
        "responses": {
          "200": {
            "description": "cluster config fields",
            "schema": {
              "$ref": "#/definitions/repairConfig"
            }
          }
        }
      },
      "put": {
        "description": "updateConfig cluster description",
        "parameters": [
          {
            "name": "conf",
            "in": "body",
            "description": "config files",
            "schema": {
              "$ref": "#/definitions/repairConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "deleteConfig cluster description",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/unit/{unit_id}/repair": {
      "put": {
        "parameters": [
          {
            "type": "string",
            "description": "cluster ID this API is performing on",
            "name": "cluster_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target unit UUID",
            "name": "unit_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "location of the new task",
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/unit/{unit_id}/stop_repair": {
      "put": {
        "parameters": [
          {
            "type": "string",
            "description": "cluster ID this API is performing on",
            "name": "cluster_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "target unit UUID",
            "name": "unit_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "location of the task",
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/unit/{unit_id}": {
      "parameters": [
        {
          "type": "string",
          "description": "cluster ID this API is performing on",
          "name": "cluster_id",
          "in": "path",
          "required": true
        },
        {
          "type": "string",
          "description": "target unit UUID",
          "name": "unit_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "description": "loadUnit Description",
        "responses": {
          "200": {
            "description": "requested unit fields",
            "schema": {
              "$ref": "#/definitions/repairUnit"
            }
          }
        }
      },
      "put": {
        "description": "updateUnit Description",
        "parameters": [
          {
            "description": "unit JSON object with updated fields",
            "name": "unitFields",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/repairUnitUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "updated unit fields",
            "schema": {
              "$ref": "#/definitions/repairUnit"
            }
          }
        }
      },
      "delete": {
        "description": "deleteUnit Description",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/units": {
      "parameters": [
        {
          "type": "string",
          "description": "cluster ID this API is performing on",
          "name": "cluster_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "description": "listUnits Description",
        "responses": {
          "200": {
            "description": "array of units",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/repairUnit"
              }
            }
          }
        }
      },
      "post": {
        "description": "createUnit Description",
        "parameters": [
          {
            "description": "unit JSON object with updated fields",
            "name": "unitFields",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/repairUnitUpdate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "location of the new unit",
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}/repair/task/{task_id}": {
      "parameters": [
        {
          "type": "string",
          "description": "cluster ID this API is performing on",
          "name": "cluster_id",
          "in": "path",
          "required": true
        },
        {
          "type": "string",
          "description": "task ID requesting the stats",
          "name": "task_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "parameters": [
          {
            "description": "unit ID requesting the stats",
            "type": "string",
            "name": "unit_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "percentage of run completed in total and details",
            "schema": {
              "$ref": "#/definitions/repairProgress"
            }
          }
        }
      }
    }
  }
}